Move activation into lifecycle.rs
